guard against audio stream corruption

+++ demodulator.go
@@ -6,14 +6,14 @@ type Demodulator interface {
 	Decoder() <-chan Message
 
 	// SampleRate is Hz.
-	SampleRate() uint32
+	SampleRate() int
 
 	// BitDepth, e.g. 16, 32.
-	BitDepth() uint16
+	BitDepth() int
 
 	// NumChannels, e.g. 1 (mono), 2 (stereo).
-	NumChannels() uint16
+	NumChannels() int
 
 	// WavCategory, e.g. 1 (PCM).
-	WavCategory() uint16
+	WavCategory() int
 }
